content/canvas/test/test_context.unrecognised.nullsuffix.html

<!DOCTYPE HTML>
<title>Canvas test: context.unrecognised.nullsuffix - bug 401788</title>
<!-- Testing: Context name "2d" plus a "\0" suffix is unrecognised -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;
try {

ok(canvas.getContext("2d\0") === null, "canvas.getContext(\"2d\\0\") === null");

} catch (e) {
    _thrown_outer = true;
}
todo(!_thrown_outer, 'should not throw exception');

SimpleTest.finish();

});
</script>
